Hitting a wall, need to determine if it's worth the efforts of making their API consumable on this end
Might just want to do whatever's easier to display on the UI of the app which would be a string - as long as it needs to be. Probably without the Footnotes too. Unless I want to parse those out like I have and build a fancy UI for that... TBD