Adds Podcast RSS feed generation service
Implements a new service to trigger Lambda functions for podcast RSS feed generation. This includes the creation of an interface and implementation for invoking Lambda functions that rebuild the feed and upsert individual episodes. The service is integrated into the Sermons service to trigger feed updates when new sermons are created or updated, and the service can be mocked/tested through DI.

*   Adds a new `IPodcastLambdaService` interface and `PodcastLambdaService` implementation.
*   Updates the `SermonsService` to use the new service when new messages are created or updated.
*   Adds tests for the new service and updates existing tests to inject the new service.
*   Updates the AWS SDK S3 version for consistency.